eleiminated extension methods by returning IEnumerable<> directly from the methods that return collections.
implemented the imanobject interface from the SDK. the me member was also declared at this level. all wrapped classes should now derive from this object and override the me object.

implemented the imanobjectdatabase class, from which all other wrapper classes that need a reference to a database object will derive from. this required editing all of the classes to remove the database member and re-do the constructors.

implemented all properties of imandocument, including read/write. also implemented a few other classes to make this happen, even if their implementation is not complete.